% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{apply_crses_to_sf_stream}
\alias{apply_crses_to_sf_stream}
\title{This is a workaround for the current availability of the GDAL stream interface
as exposed by the sf package. With}
\usage{
apply_crses_to_sf_stream(
  stream_in_xptr,
  geometry_column_names,
  geometry_column_crses,
  stream_out_xptr
)
}
\description{
This is a workaround for the current availability of the GDAL stream interface
as exposed by the sf package. With
}
