% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as_sedonadb_dataframe}
\alias{as_sedonadb_dataframe}
\title{Convert an object to a DataFrame}
\usage{
as_sedonadb_dataframe(x, ..., schema = NULL, ctx = NULL)
}
\arguments{
\item{x}{An object to convert}

\item{...}{Extra arguments passed to/from methods}

\item{schema}{The requested schema}

\item{ctx}{A SedonaDB context. This should always be passed to inner calls
to SedonaDB functions; NULL implies the global context.}
}
\value{
A sedonadb_dataframe
}
\description{
Convert an object to a DataFrame
}
\examples{
as_sedonadb_dataframe(data.frame(x = 1:3))

}
