% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{as_sedonadb_expr}
\alias{as_sedonadb_expr}
\title{Create a SedonaDB Logical Expression}
\usage{
as_sedonadb_expr(x, ..., type = NULL)
}
\arguments{
\item{x}{An object}

\item{...}{Passed to/from methods}

\item{type}{An optional data type to request for the output}
}
\value{
An object of class SedonaDBExpr
}
\description{
Create a SedonaDB Logical Expression
}
