% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literal.R
\name{as_sedonadb_literal}
\alias{as_sedonadb_literal}
\title{S3 Generic to create a SedonaDB literal expression}
\usage{
as_sedonadb_literal(x, ..., type = NULL, factory = NULL)
}
\arguments{
\item{x}{An object to convert to a SedonaDB literal}

\item{...}{Passed to/from methods}

\item{type}{An optional data type to request for the output}

\item{factory}{An \code{sd_expr_factory()} that should be passed to any
other calls to \code{as_sedonadb_literal()} if needed}
}
\value{
An object of class SedonaDBExpr
}
\description{
This generic provides the opportunity for objects to register a mechanism
to be understood as literals in the context of a SedonaDB expression.
Users constructing expressions directly should use \code{\link[=sd_expr_literal]{sd_expr_literal()}}.
}
\examples{
as_sedonadb_literal("abcd")

}
