% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{.fns}
\alias{.fns}
\title{SedonaDB Functions}
\usage{
.fns
}
\description{
This object is an escape hatch for calling SedonaDB/DataFusion functions
directly for translations that are not yet registered or are otherwise
misbehaving.
}
