% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-expression.R
\docType{data}
\name{.tables}
\alias{.tables}
\title{SedonaDB Table Qualifiers}
\format{
An object of class \code{sedonadb_tables} of length 2.
}
\usage{
.tables
}
\description{
This object is an escape hatch for referring to the right or left side
of a join when constructing \code{\link[=sd_join_by]{sd_join_by()}} or \code{\link[=sd_join_select]{sd_join_select()}} expressions.
}
\keyword{datasets}
