% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_arrange}
\alias{sd_arrange}
\title{Order rows of a SedonaDB data frame using column values}
\usage{
sd_arrange(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe or an object that can be coerced to one.}

\item{...}{Unnamed expressions for filter conditions. These are evaluated
in the same way as \code{\link[dplyr:arrange]{dplyr::arrange()}} except does not support extra
dplyr features such as \code{across()}, \code{.by_group}, or \code{.locale}.}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Order rows of a SedonaDB data frame using column values
}
\examples{
data.frame(x = c(10:1, NA)) |> sd_arrange(x)
data.frame(x = c(10:1, NA)) |> sd_arrange(desc(x))

}
