% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_compute}
\alias{sd_compute}
\alias{sd_collect}
\title{Collect a DataFrame into memory}
\usage{
sd_compute(.data)

sd_collect(.data, ptype = NULL)
}
\arguments{
\item{.data}{A sedonadb_dataframe}

\item{ptype}{The target R object. See \link[nanoarrow:convert_array_stream]{nanoarrow::convert_array_stream}.}
}
\value{
\code{sd_compute()} returns a sedonadb_dataframe; \code{sd_collect()} returns
a data.frame (or subclass according to \code{ptype}).
}
\description{
Use \code{sd_compute()} to collect and return the result as a DataFrame;
use \code{sd_collect()} to collect and return the result as an R data.frame.
}
\examples{
sd_sql("SELECT 1 as one") |> sd_compute()
sd_sql("SELECT 1 as one") |> sd_collect()

}
