% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_count}
\alias{sd_count}
\title{Count rows in a DataFrame}
\usage{
sd_count(.data)
}
\arguments{
\item{.data}{A sedonadb_dataframe}
}
\value{
The number of rows after executing the query
}
\description{
Count rows in a DataFrame
}
\examples{
sd_sql("SELECT 1 as one") |> sd_count()

}
