% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sd_drop_view}
\alias{sd_drop_view}
\alias{sd_view}
\title{Create or Drop a named view}
\usage{
sd_drop_view(table_ref)

sd_view(table_ref)
}
\arguments{
\item{table_ref}{The name of the view reference}
}
\value{
The context, invisibly
}
\description{
Remove a view created with \code{\link[=sd_to_view]{sd_to_view()}} from the context.
}
\examples{
sd_sql("SELECT 1 as one") |> sd_to_view("foofy")
sd_view("foofy")
sd_drop_view("foofy")
try(sd_view("foofy"))

}
