% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_filter}
\alias{sd_filter}
\title{Keep rows of a SedonaDB DataFrame that match a condition}
\usage{
sd_filter(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe or an object that can be coerced to one.}

\item{...}{Unnamed expressions for filter conditions. These are evaluated
in the same way as \code{\link[dplyr:filter]{dplyr::filter()}} except does not support extra
dplyr features such as \code{across()} or \code{.by}.}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Keep rows of a SedonaDB DataFrame that match a condition
}
\examples{
data.frame(x = 1:10) |> sd_filter(x > 5)

}
