% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_group_by}
\alias{sd_group_by}
\title{Group SedonaDB DataFrames by one or more expressions}
\usage{
sd_group_by(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe or an object that can be coerced to one.}

\item{...}{Named expressions whose unique combination will be used as
groups to potentially compute a future aggregate expression. These are
evaluated in the same way as \code{\link[dplyr:group_by]{dplyr::group_by()}} except \code{.add} nor
\code{.drop} are supported.}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Note that unlike \code{\link[dplyr:group_by]{dplyr::group_by()}}, these groups are dropped after
any transformations.
}
\examples{
data.frame(letter = c(rep("a", 3, rep("b", 4), rep("c", 3))) x = 1:10) |>
  sd_group_by(letter) |>
  sd_summarise(x = sum(x))

}
