% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-expression.R
\name{sd_join_by}
\alias{sd_join_by}
\alias{sd_join_intersects}
\alias{sd_join_contains}
\alias{sd_join_within}
\alias{sd_join_covers}
\alias{sd_join_coveredby}
\alias{sd_join_touches}
\alias{sd_join_crosses}
\alias{sd_join_overlaps}
\alias{sd_join_equals}
\alias{sd_join_dwithin}
\title{Specify join conditions}
\usage{
sd_join_by(...)

sd_join_intersects()

sd_join_contains()

sd_join_within()

sd_join_covers()

sd_join_coveredby()

sd_join_touches()

sd_join_crosses()

sd_join_overlaps()

sd_join_equals()

sd_join_dwithin(distance)
}
\arguments{
\item{...}{Expressions specifying join conditions. These should be
comparison expressions (e.g., \code{x$id == y$id}, \code{x$value > y$threshold})
or spatial predicate expressions
(e.g., \code{st_intersects(x$geometry, y$geometry)}).
Multiple conditions are combined with AND. Like dplyr's \code{join_by()},
single columns are parsed as an equijoin condition (e.g., \code{id} becomes
\code{x$id == y$id}).}

\item{distance}{For a within-distance join, the distance threshold.}
}
\value{
An object of class \code{sedonadb_join_by} containing the unevaluated
join condition expressions.
}
\description{
Use \code{sd_join_by()} to specify join conditions for \code{\link[=sd_join]{sd_join()}} using
expressions that reference columns from both tables. Table references
are specified using \code{x$column} and \code{y$column} syntax to disambiguate
columns from the left and right tables, and the special helper \code{x$geom()}
and \code{y$geom()} may be used for tables with exactly one geometry column.
Spatial joins can use spatial predicates in the join by expression
(e.g., \code{sd_join_by(st_intersects(x$geom(), y$geom()))}) or the shorthand
\code{sd_join_intersects()}.
}
\details{
For programmatic usage, the \code{.tables} pronoun may be used to unambiguously
refer to a table qualifier (similar to the \code{.data} pronoun which may be
used in single-table SedonaDB verbs).
}
\examples{
# Equality join on id column
sd_join_by(x$id == y$id)

# Can use just the column name as a shorthand
sd_join_by(id)

# Multiple conditions (combined with AND)
sd_join_by(x$id == y$id, x$date >= y$start_date)

# Inequality join
sd_join_by(x$value > y$threshold)

# Spatial joins
sd_join_intersects()
sd_join_dwithin(100)

}
