% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_preview}
\alias{sd_preview}
\title{Preview and print the results of running a query}
\usage{
sd_preview(.data, n = NULL, ascii = NULL, width = NULL)
}
\arguments{
\item{.data}{A sedonadb_dataframe}

\item{n}{The number of rows to preview. Use \code{Inf} to preview all rows.
Defaults to \code{getOption("pillar.print_max")}.}

\item{ascii}{Use \code{TRUE} to force ASCII table formatting or \code{FALSE} to force
unicode formatting. By default, use a heuristic to determine if the output
is unicode-friendly or the value of \code{getOption("cli.unicode")}.}

\item{width}{The character width of the output. Defaults to
\code{getOption("width")}.}
}
\value{
.data, invisibly
}
\description{
This is used to implement \code{print()} for the sedonadb_dataframe or can
be used to explicitly preview if \code{options(sedonadb.interactive = FALSE)}.
}
\examples{
sd_sql("SELECT 1 as one") |> sd_preview()

}
