% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{sd_read_sf}
\alias{sd_read_sf}
\alias{sd_ctx_read_sf}
\title{Read GDAL/OGR via the sf package}
\usage{
sd_read_sf(
  dsn,
  layer = NULL,
  ...,
  query = NA,
  options = NULL,
  drivers = NULL,
  filter = NULL,
  fid_column_name = NULL,
  lazy = FALSE
)

sd_ctx_read_sf(
  ctx,
  dsn,
  layer = NULL,
  ...,
  query = NA,
  options = NULL,
  drivers = NULL,
  filter = NULL,
  fid_column_name = NULL,
  lazy = FALSE
)
}
\arguments{
\item{dsn, layer}{Description of datasource and layer. See \code{\link[sf:read_sf]{sf::read_sf()}}
for details.}

\item{...}{Currently unused and must be empty}

\item{query}{A SQL query to pass on to GDAL/OGR.}

\item{options}{A character vector with layer open options in the
form "KEY=VALUE".}

\item{drivers}{A list of drivers to try if the dsn cannot be guessed.}

\item{filter}{A spatial object that may be used to filter while reading.
In the future SedonaDB will automatically calculate this value based on
the query. May be any spatial object that can be converted to WKT via
\code{\link[wk:as_wkt]{wk::as_wkt()}}. This filter's CRS must match that of the data.}

\item{fid_column_name}{An optional name for the feature id (FID) column.}

\item{lazy}{Use \code{TRUE} to stream the data from the source rather than collect
first. This can be faster for large data sources but can also be confusing
because the data may only be scanned exactly once.}

\item{ctx}{A SedonaDB context created using \code{\link[=sd_connect]{sd_connect()}}.}
}
\value{
A SedonaDB DataFrame.
}
\description{
Uses the ArrowArrayStream interface to GDAL exposed via the sf package
to read GDAL/OGR-based data sources.
}
\examples{
nc_gpkg <- system.file("gpkg/nc.gpkg", package = "sf")
sd_read_sf(nc_gpkg)

}
