% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sd_register_udf}
\alias{sd_register_udf}
\title{Register a user-defined function}
\usage{
sd_register_udf(udf)
}
\arguments{
\item{udf}{An object of class 'datafusion_scalar_udf'}
}
\value{
NULL, invisibly
}
\description{
Register a user-defined function
}
