% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_select}
\alias{sd_select}
\title{Keep or drop columns of a SedonaDB DataFrame}
\usage{
sd_select(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe}

\item{...}{One or more bare names. Evaluated like \code{\link[dplyr:select]{dplyr::select()}}.}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Keep or drop columns of a SedonaDB DataFrame
}
\examples{
data.frame(x = 1:10, y = letters[1:10]) |> sd_select(x)

}
