% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sd_sql}
\alias{sd_sql}
\alias{sd_ctx_sql}
\title{Create a DataFrame from SQL}
\usage{
sd_sql(sql, ..., params = NULL)

sd_ctx_sql(ctx, sql, ..., params = NULL)
}
\arguments{
\item{sql}{A SQL string to execute}

\item{...}{These dots are for future extensions and currently must be empty.}

\item{params}{A list of parameters to fill placeholders in the query.}

\item{ctx}{A SedonaDB context.}
}
\value{
A sedonadb_dataframe
}
\description{
The query will only be executed when requested.
}
\examples{
sd_sql("SELECT ST_Point(0, 1) as geom")
sd_sql("SELECT ST_Point($1, $2) as geom", params = list(1, 2))
sd_sql("SELECT ST_Point($x, $y) as geom", params = list(x = 1, y = 2))

}
