% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_to_view}
\alias{sd_to_view}
\title{Register a DataFrame as a named view}
\usage{
sd_to_view(.data, table_ref, overwrite = FALSE)
}
\arguments{
\item{.data}{A sedonadb_dataframe}

\item{table_ref}{The name of the view reference}

\item{overwrite}{Use TRUE to overwrite a view with the same name (if it exists)}
}
\value{
.data, invisibly
}
\description{
This is useful for creating a view that can be referenced in a SQL
statement. Use \code{\link[=sd_drop_view]{sd_drop_view()}} to remove it.
}
\examples{
sd_sql("SELECT 1 as one") |> sd_to_view("foofy")
sd_sql("SELECT * FROM foofy")

}
