% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_transmute}
\alias{sd_transmute}
\title{Create, modify, and delete columns of a SedonaDB DataFrame}
\usage{
sd_transmute(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe}

\item{...}{Named expressions for new columns to create. These are evaluated
in the same way as \code{\link[dplyr:transmute]{dplyr::transmute()}} except does not support extra
dplyr features such as \code{across()} or \code{.by}.}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Create, modify, and delete columns of a SedonaDB DataFrame
}
\examples{
data.frame(x = 1:10) |>
  sd_transmute(y = x + 1L)

}
