% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_with_params}
\alias{sd_with_params}
\title{Fill in placeholders}
\usage{
sd_with_params(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe or an object that can be coerced to one.}

\item{...}{Named or unnamed parameters that will be coerced to literals
with \code{\link[=as_sedonadb_literal]{as_sedonadb_literal()}}.}
}
\value{
A sedonadb_dataframe with the provided parameters filled into the query
}
\description{
This is a slightly more verbose form of \code{\link[=sd_sql]{sd_sql()}} with \code{params} that is
useful if a data frame is to be repeatedly queried.
}
\examples{
sd_sql("SELECT ST_Point($1, $2) as pt") |>
  sd_with_params(11, 12)
sd_sql("SELECT ST_Point($x, $y) as pt") |>
  sd_with_params(x = 11, y = 12)

}
