% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedonadb-package.R
\docType{package}
\name{sedonadb-package}
\alias{sedonadb}
\alias{sedonadb-package}
\title{sedonadb: Bindings for Apache SedonaDB}
\description{
Provides bindings for Apache SedonaDB, a lightweight query engine optimized for spatial workflows.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@dunnington.ca}

}
\keyword{internal}
