% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adbc.R
\name{sedonadb_adbc}
\alias{sedonadb_adbc}
\title{SedonaDB ADBC Driver}
\usage{
sedonadb_adbc()
}
\value{
An \code{\link[adbcdrivermanager:adbc_driver_void]{adbcdrivermanager::adbc_driver()}} of class
'sedonadb_driver_sedonadb'
}
\description{
SedonaDB ADBC Driver
}
\examples{
library(adbcdrivermanager)

con <- sedonadb_adbc() |>
  adbc_database_init() |>
  adbc_connection_init()
con |>
  read_adbc("SELECT ST_Point(0, 1) as geometry") |>
  as.data.frame()

}
