% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{as_sedonadb_literal}
\alias{as_sedonadb_literal}
\title{Create a SedonaDB Literal Expression}
\usage{
as_sedonadb_literal(x, ..., type = NULL, factory = NULL)
}
\arguments{
\item{x}{An object to convert to a SedonaDB literal}

\item{...}{Passed to/from methods}

\item{type}{An optional data type to request for the output}

\item{factory}{An expression factory object that should be passed to any
other calls to \code{as_sedonadb_literal()}.}
}
\value{
An object of class SedonaDBExpr
}
\description{
Create a SedonaDB Literal Expression
}
