% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{sd_expr_column}
\alias{sd_expr_column}
\alias{sd_expr_literal}
\alias{sd_expr_scalar_function}
\alias{as_sd_expr}
\alias{is_sd_expr}
\alias{sd_expr_factory}
\title{Create SedonaDB logical expressions}
\usage{
sd_expr_column(column_name, qualifier = NULL, ..., factory = sd_expr_factory())

sd_expr_literal(x, ..., type = NULL, factory = sd_expr_factory())

sd_expr_scalar_function(function_name, args, ..., factory = sd_expr_factory())

as_sd_expr(x, ..., factory = sd_expr_factory())

is_sd_expr(x)

sd_expr_factory()
}
\arguments{
\item{column_name}{A column name}

\item{qualifier}{An optional qualifier (e.g., table reference) that may be
used to disambiguate a specific reference}

\item{...}{Used to accommodate future expansion of the API. Currently all
dots must be empty.}

\item{factory}{A \code{\link[=sd_expr_factory]{sd_expr_factory()}}. This factory wraps a SedonaDB context
and is used to resolve scalar functions and/or retrieve options.}
}
\value{
An object of class SedonaDBExpr
}
\description{
Create SedonaDB logical expressions
}
\examples{
sd_expr_column("foofy")
sd_expr_literal(1L)
sd_expr_scalar_function("abs", list(1L))

}
