% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_join}
\alias{sd_join}
\title{Join two SedonaDB DataFrames}
\usage{
sd_join(x, y, by = NULL, suffix = c(".x", ".y"), join_type = "inner")
}
\arguments{
\item{x}{The left dataframe}

\item{y}{The right dataframe (will use the same context as x)}

\item{by}{A \code{sedonadb_join_by} object from \code{\link[=sd_join_by]{sd_join_by()}}, or \code{NULL} for
a natural join on columns with matching names.}

\item{suffix}{A character vector of length 2 specifying suffixes to add
to overlapping column names from x and y respectively.}

\item{how}{The type of join to perform. One of "inner", "left", "right",
"full", "leftsemi", "rightsemi", "leftanti", "rightanti", "leftmark",
or "rightmark".}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Perform a join operation between two dataframes. Use \code{\link[=sd_join_by]{sd_join_by()}} to
specify join conditions using \code{x$column} and \code{y$column} syntax to
reference columns from the left and right tables respectively.
}
\examples{
library(sedonadb)

df1 <- data.frame(x = letters[1:10], y = 1:10)
df2 <- data.frame(y = 10:1, z = stringr::words[1:10])
df1 |> sd_join(df2)

}
