% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_join}
\alias{sd_join}
\title{Join two SedonaDB DataFrames}
\usage{
sd_join(
  x,
  y,
  by = NULL,
  join_type = "inner",
  select = sd_join_select_default()
)
}
\arguments{
\item{x}{The left dataframe}

\item{y}{The right dataframe (will use the same context as x)}

\item{by}{Join specification. One of:
\itemize{
\item A \code{sedonadb_join_by} object from \code{\link[=sd_join_by]{sd_join_by()}}
\item A character vector of column names to join on in both tables
\item A named character vector mapping left-table column names to
right-table column names, e.g. \code{c(x_val = "y_val")}
\item \code{NULL} for a natural join on columns with matching names
}}

\item{join_type}{The type of join to perform. One of "inner", "left", "right",
"full", "leftsemi", "rightsemi", "leftanti", "rightanti", "leftmark",
or "rightmark".}

\item{select}{Post-join column selection. One of
\itemize{
\item \code{NULL} for no modification, which may result in duplicate (unqualified)
column names. The column may still be
referred to with a qualifier in advanced usage using \code{\link[=sd_expr_column]{sd_expr_column()}}.
\item \code{\link[=sd_join_select_default]{sd_join_select_default()}} for dplyr-like behaviour (equi-join keys
removed, intersecting names suffixed)
\item \code{\link[=sd_join_select]{sd_join_select()}} for a custom selection
}}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Perform a join operation between two dataframes. Use \code{\link[=sd_join_by]{sd_join_by()}} to
specify join conditions using \code{x$column} and \code{y$column} syntax to
reference columns from the left and right tables respectively.
}
\examples{
df1 <- data.frame(x = letters[1:10], y = 1:10)
df2 <- data.frame(y = 10:1, z = LETTERS[1:10])
df1 |> sd_join(df2)

}
