% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-expression.R
\name{sd_join_by}
\alias{sd_join_by}
\title{Specify join conditions}
\usage{
sd_join_by(...)
}
\arguments{
\item{...}{Expressions specifying join conditions. These should be
comparison expressions (e.g., \code{x$id == y$id}, \code{x$value > y$threshold})
or spatial predicate expressions
(e.g., \code{st_intersects(x$geometry, y$geometry)}).
Multiple conditions are combined with AND.}
}
\value{
An object of class \code{sedonadb_join_by} containing the unevaluated
join condition expressions.
}
\description{
Use \code{sd_join_by()} to specify join conditions for \code{\link[=sd_join]{sd_join()}} using
expressions that reference columns from both tables. Table references
are specified using \code{x$column} and \code{y$column} syntax to disambiguate
columns from the left and right tables.
}
\examples{
# Equality join on id column
sd_join_by(x$id == y$id)

# Multiple conditions (combined with AND)
sd_join_by(x$id == y$id, x$date >= y$start_date)

# Inequality join
sd_join_by(x$value > y$threshold)

}
