% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sd_read_parquet}
\alias{sd_read_parquet}
\title{Create a DataFrame from one or more Parquet files}
\usage{
sd_read_parquet(path)
}
\arguments{
\item{path}{One or more paths or URIs to Parquet files}
}
\value{
A sedonadb_dataframe
}
\description{
The query will only be executed when requested.
}
\examples{
path <- system.file("files/natural-earth_cities_geo.parquet", package = "sedonadb")
sd_read_parquet(path) |> head(5) |> sd_preview()

}
