% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sd_register_udf}
\alias{sd_register_udf}
\title{Register a user-defined function}
\usage{
sd_register_udf(udf)
}
\arguments{
\item{udf}{An object of class 'datafusion_scalar_udf'}
}
\value{
NULL, invisibly
}
\description{
Several types of user-defined functions can be registered into a session
context. Currently, the only implemented variety is an external pointer
to a Rust \code{FFI_ScalarUDF}, an example of which is available from the
\href{https://github.com/apache/datafusion-python/blob/6f3b1cab75cfaa0cdf914f9b6fa023cb9afccd7d/examples/datafusion-ffi-example/src/scalar_udf.rs}{DataFusion Python documentation}.
}
