% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{sd_sql}
\alias{sd_sql}
\title{Create a DataFrame from SQL}
\usage{
sd_sql(sql)
}
\arguments{
\item{sql}{A SQL string to execute}
}
\value{
A sedonadb_dataframe
}
\description{
The query will only be executed when requested.
}
\examples{
sd_sql("SELECT ST_Point(0, 1) as geom") |> sd_preview()

}
