% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{sd_summarise}
\alias{sd_summarise}
\alias{sd_summarize}
\title{Order rows of a SedonaDB data frame using column values}
\usage{
sd_summarise(.data, ...)

sd_summarize(.data, ...)
}
\arguments{
\item{.data}{A sedonadb_dataframe or an object that can be coerced to one.}

\item{...}{Aggregate expressions. These are evaluated in the same way as
\code{\link[dplyr:summarise]{dplyr::summarise()}} except the outer expression must be an aggregate
expression (e.g., \code{sum(x) + 1} is not currently possible).}
}
\value{
An object of class sedonadb_dataframe
}
\description{
Order rows of a SedonaDB data frame using column values
}
\examples{
data.frame(x = c(10:1, NA)) |> sd_summarise(x = sum(x, na.rm = TRUE))

}
